+++ mock-abis/src/main/java/io/mosip/proxy/abis/controller/ProxyAbisConfigController.java
@@ -1,34 +1,21 @@
 package io.mosip.proxy.abis.controller;
 
-import javax.validation.Valid;
-
+import io.mosip.proxy.abis.entity.ConfigureDto;
+import io.mosip.proxy.abis.entity.Expectation;
 import io.mosip.proxy.abis.service.ProxyAbisInsertService;
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiOperation;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.bind.annotation.*;
 
-import io.mosip.proxy.abis.entity.Expectation;
-import io.mosip.proxy.abis.entity.FailureResponse;
-import io.mosip.proxy.abis.entity.IdentityRequest;
-import io.mosip.proxy.abis.entity.IdentityResponse;
-import io.mosip.proxy.abis.entity.InsertRequestMO;
-import io.mosip.proxy.abis.entity.RequestMO;
-import io.mosip.proxy.abis.entity.ResponseMO;
-import io.mosip.proxy.abis.exception.BindingException;
-import io.mosip.proxy.abis.exception.FailureReasonsConstants;
-import io.mosip.proxy.abis.exception.RequestException;
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
+import javax.validation.Valid;
+import java.util.Map;
+
 
 @CrossOrigin
 @RestController
@@ -43,161 +30,69 @@ public class ProxyAbisConfigController {
 
     @RequestMapping(value = "expectation", method = RequestMethod.POST)
     @ApiOperation(value = "Sets expectation")
-    public ResponseEntity<Boolean> setExpectation(@Valid @RequestBody Expectation expectation) throws Exception {
+    public ResponseEntity<String> setExpectation(@Valid @RequestBody Expectation expectation) throws Exception {
         logger.info("Setting expectation" + expectation.getId());
-
+        try {
+            abisInsertService.setExpectation(expectation);
+            return new ResponseEntity<>("Successfully inserted expectation "+expectation.getId(), HttpStatus.OK);
+        } catch (RuntimeException exp) {
+            logger.error("Exception while getting expectation: "+exp.getMessage());
+            throw exp;
+        }
     }
 
     @RequestMapping(value = "expectation", method = RequestMethod.GET)
     @ApiOperation(value = "Gets expectation")
-    public ResponseEntity<Expectation> getExpectation() throws Exception {
+    public ResponseEntity<Map<String, Expectation>> getExpectation() throws Exception {
         logger.info("Getting expectation");
-
-
+        try {
+            return new ResponseEntity<>(abisInsertService.getExpectations(), HttpStatus.OK);
+        } catch (RuntimeException exp) {
+            logger.error("Exception while getting expectation: "+exp.getMessage());
+            throw exp;
+        }
     }
 
-    public ResponseEntity<Object> saveInsertRequest(@Valid @RequestBody InsertRequestMO ie, BindingResult bd)
+    @RequestMapping(value = "expectation/{id}", method = RequestMethod.DELETE)
+    @ApiOperation(value = "Delete expectation")
+    public ResponseEntity<String> deleteExpectation(@PathVariable String id) {
+        logger.info("Delete expectation: "+id);
+        try {
+            abisInsertService.deleteExpectation(id);
+            return new ResponseEntity<>("Successfully deleted expectation "+id, HttpStatus.OK);
+        } catch (RuntimeException exp) {
+            logger.error("Exception while getting expectation: "+exp.getMessage());
+            throw exp;
+        }
+    }
+
+    @RequestMapping(value = "configure", method = RequestMethod.GET)
+    @ApiOperation(value = "Configure Request")
+    public ResponseEntity<ConfigureDto> checkConfiguration()
             throws Exception {
-        logger.info("Saving Insert Request");
-        if (bd.hasErrors()) {
-            logger.info("Some fields are missing in the insert request");
-            RequestMO re = new RequestMO(ie.getId(), ie.getVersion(), ie.getRequestId(), ie.getRequesttime(),
-                    ie.getReferenceId());
-            throw new BindingException(re, bd);
-        }
+        logger.info("Configure Request");
         try {
-            return processInsertRequest(ie);
-        } catch (RequestException exp) {
-            logger.error("Exception while saving insert request");
-            RequestMO re = new RequestMO(ie.getId(), ie.getVersion(), ie.getRequestId(), ie.getRequesttime(),
-                    ie.getReferenceId());
-            exp.setEntity(re);
-            if (null == exp.getReasonConstant())
-                exp.setReasonConstant(FailureReasonsConstants.INTERNAL_ERROR_UNKNOWN);
+            ConfigureDto configureDto = new ConfigureDto();
+            configureDto.setFindDuplicate(abisInsertService.getDuplicate());
+            return new ResponseEntity<>(configureDto, HttpStatus.OK);
+        } catch (RuntimeException exp) {
+            logger.error("Exception while getting configuration: "+exp.getMessage());
             throw exp;
         }
     }
 
-    @RequestMapping(value = "deleterequest", method = RequestMethod.DELETE)
-    @ApiOperation(value = "Delete Request")
-    public ResponseEntity<Object> deleteRequest(@RequestBody RequestMO ie) {
+    @RequestMapping(value = "configure", method = RequestMethod.POST)
+    @ApiOperation(value = "Configure Request")
+    public ResponseEntity<String> configure(@Valid @RequestBody ConfigureDto ie, BindingResult bd)
+            throws Exception {
+        logger.info("Configure Request");
         try {
-            return processDeleteRequest(ie);
-        } catch (RequestException exp) {
-            logger.error("Exception while deleting reference id");
-            exp.setEntity(ie);
-            exp.setReasonConstant(FailureReasonsConstants.INTERNAL_ERROR_UNKNOWN);
+            abisInsertService.setDuplicate(ie.getFindDuplicate());
+            logger.info("[Configuration updated] overrideFindDuplicate: "+abisInsertService.getDuplicate());
+            return new ResponseEntity<>("Successfully updated the configuration", HttpStatus.OK);
+        } catch (RuntimeException exp) {
+            logger.error("Exception in configure request: "+exp.getMessage());
             throw exp;
         }
-
     }
-
-    @RequestMapping(value = "upload", method = RequestMethod.POST)
-    @ApiOperation(value = "Upload certificate Request")
-    public ResponseEntity<String> uploadcertificate(@RequestParam("file") MultipartFile uploadfile,
-                                                    @RequestParam("password") String password, @RequestParam("alias") String alias,@RequestParam("keystore") String keystore) {
-        if (uploadfile.isEmpty())
-            return new ResponseEntity("Please select a file", HttpStatus.NO_CONTENT);
-
-        if (null == alias || alias.isEmpty())
-            return new ResponseEntity("Please enter alias", HttpStatus.NO_CONTENT);
-        if (null == password || password.isEmpty())
-            return new ResponseEntity("Please enter password", HttpStatus.NO_CONTENT);
-
-        return new ResponseEntity<String>( abisInsertService.saveUploadedFileWithParameters(uploadfile, alias, password,keystore),HttpStatus.OK);
-
-
-    }
-
-    @RequestMapping(value = "identityrequest", method = RequestMethod.POST)
-    @ApiOperation(value = "Checks duplication")
-    public ResponseEntity<Object> identityRequest(@RequestBody IdentityRequest ir) {
-        try {
-            return processIdentityRequest(ir);
-        } catch (Exception ex) {
-            logger.error("Error while finding duplicates for " + ir.getReferenceId());
-            RequestMO re = new RequestMO(ir.getId(), ir.getVersion(), ir.getRequestId(), ir.getRequesttime(),
-                    ir.getReferenceId());
-            throw new RequestException(re, FailureReasonsConstants.UNABLE_TO_FETCH_BIOMETRIC_DETAILS);
-        }
-    }
-
-    public ResponseEntity<Object> deleteRequestThroughListner(RequestMO ie) {
-        try {
-            return processDeleteRequest(ie);
-        } catch (Exception ex) {
-            FailureResponse fr = new FailureResponse(ie.getId(), ie.getRequestId(), ie.getRequesttime(), 2,
-                    FailureReasonsConstants.INTERNAL_ERROR_UNKNOWN);
-            return new ResponseEntity<Object>(fr, HttpStatus.NOT_ACCEPTABLE);
-        }
-    }
-
-    private ResponseEntity<Object> processDeleteRequest(RequestMO ie) {
-        logger.info("Deleting request with reference id" + ie.getReferenceId());
-        abisInsertService.deleteData(ie.getReferenceId());
-        ResponseMO response = new ResponseMO(ie.getId(), ie.getRequestId(), ie.getRequesttime(), 1);
-        logger.info("Successfully deleted reference id" + ie.getReferenceId());
-        return new ResponseEntity<Object>(response, HttpStatus.OK);
-    }
-
-    public ResponseEntity<Object> identityRequestThroughListner(IdentityRequest ir) {
-        try {
-            return processIdentityRequest(ir);
-        } catch (Exception ex) {
-            FailureResponse fr = new FailureResponse(ir.getId(), ir.getRequestId(), ir.getRequesttime(), 2,
-                    FailureReasonsConstants.UNABLE_TO_FETCH_BIOMETRIC_DETAILS);
-            return new ResponseEntity<Object>(fr, HttpStatus.NOT_ACCEPTABLE);
-        }
-    }
-
-    private ResponseEntity<Object> processIdentityRequest(IdentityRequest ir) {
-        logger.info("Finding duplication for reference ID " + ir.getReferenceId());
-        IdentityResponse res = abisInsertService.findDupication(ir);
-        return new ResponseEntity<Object>(res, HttpStatus.OK);
-    }
-
-    public ResponseEntity<Object> saveInsertRequestThroughListner(InsertRequestMO ie) {
-        logger.info("Saving Insert Request");
-        String validate = validateRequest(ie);
-        if (null != validate) {
-            FailureResponse fr = new FailureResponse(ie.getId(), ie.getRequestId(), ie.getRequesttime(), 2, validate);
-            return new ResponseEntity<Object>(fr, HttpStatus.NOT_ACCEPTABLE);
-        }
-
-        try {
-            return processInsertRequest(ie);
-        } catch (RequestException exp) {
-            FailureResponse fr = new FailureResponse(ie.getId(), ie.getRequestId(), ie.getRequesttime(), 2,
-                    null == exp.getReasonConstant() ? FailureReasonsConstants.INTERNAL_ERROR_UNKNOWN
-                            : exp.getReasonConstant().toString());
-            return new ResponseEntity<Object>(fr, HttpStatus.NOT_ACCEPTABLE);
-        }
-
-    }
-
-    private ResponseEntity<Object> processInsertRequest(InsertRequestMO ie) {
-        abisInsertService.insertData(ie);
-        ResponseMO response = new ResponseMO(ie.getId(), ie.getRequestId(), ie.getRequesttime(), 1);
-        logger.info("Successfully inserted record ");
-        return new ResponseEntity<Object>(response, HttpStatus.OK);
-    }
-
-    private String validateRequest(InsertRequestMO ie) {
-
-        if (null != ie.getId() && !ie.getId().isEmpty() && !ie.getId().equalsIgnoreCase("mosip.abis.insert"))
-            return FailureReasonsConstants.INVALID_ID;
-        if (null == ie.getRequestId() || ie.getRequestId().isEmpty())
-            return FailureReasonsConstants.MISSING_REQUESTID;
-        if (null == ie.getRequesttime())
-            return FailureReasonsConstants.MISSING_REQUESTTIME;
-        if (null == ie.getReferenceId() || ie.getReferenceId().isEmpty())
-            return FailureReasonsConstants.MISSING_REFERENCEID;
-        if (null != ie.getVersion() && !ie.getVersion().isEmpty()) {
-            if (!ie.getVersion().matches("[0-9]+.[0-9]"))
-                return FailureReasonsConstants.INVALID_VERSION;
-        }
-        return null;
-
-    }
-
 }
