+++ mock-abis/src/main/java/io/mosip/proxy/abis/controller/ProxyAbisController.java
@@ -1,14 +1,5 @@
 package io.mosip.proxy.abis.controller;
 
-import java.io.File;
-import java.io.FileWriter;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
 import javax.validation.Valid;
 
 import io.mosip.proxy.abis.entity.*;
@@ -25,7 +16,6 @@ import io.mosip.proxy.abis.exception.BindingException;
 import io.mosip.proxy.abis.exception.FailureReasonsConstants;
 import io.mosip.proxy.abis.exception.RequestException;
 import io.mosip.proxy.abis.service.ProxyAbisInsertService;
-import io.mosip.proxy.abis.service.impl.ProxyAbisInsertServiceImpl;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 
@@ -39,74 +29,6 @@ public class ProxyAbisController {
 
 	@Autowired
 	ProxyAbisInsertService abisInsertService;
-
-	@RequestMapping(value = "expectation", method = RequestMethod.POST)
-	@ApiOperation(value = "Sets expectation")
-	public ResponseEntity<Expectation> setExpectation(@Valid @RequestBody Expectation expectation) throws Exception {
-		logger.info("Setting expectation" + expectation.getId());
-		try {
-			abisInsertService.setExpectation(expectation);
-			return new ResponseEntity<>(expectation, HttpStatus.OK);
-		} catch (RuntimeException exp) {
-			logger.error("Exception while getting expectation: "+exp.getMessage());
-			throw exp;
-		}
-	}
-
-	@RequestMapping(value = "expectation", method = RequestMethod.GET)
-	@ApiOperation(value = "Gets expectation")
-	public ResponseEntity<Map<String, Expectation>> getExpectation() throws Exception {
-		logger.info("Getting expectation");
-		try {
-			return new ResponseEntity<>(abisInsertService.getExpectations(), HttpStatus.OK);
-		} catch (RuntimeException exp) {
-			logger.error("Exception while getting expectation: "+exp.getMessage());
-			throw exp;
-		}
-	}
-
-	@RequestMapping(value = "expectation/{id}", method = RequestMethod.DELETE)
-	@ApiOperation(value = "Delete expectation")
-	public ResponseEntity<String> deleteExpectation(@PathVariable String id) {
-		logger.info("Delete expectation: "+id);
-		try {
-			abisInsertService.deleteExpectation(id);
-			return new ResponseEntity<>("Successfully deleted expectation "+id, HttpStatus.OK);
-		} catch (RuntimeException exp) {
-			logger.error("Exception while getting expectation: "+exp.getMessage());
-			throw exp;
-		}
-	}
-
-	@RequestMapping(value = "configure", method = RequestMethod.GET)
-	@ApiOperation(value = "Configure Request")
-	public ResponseEntity<ConfigureDto> checkConfiguration()
-			throws Exception {
-		logger.info("Configure Request");
-		try {
-			ConfigureDto configureDto = new ConfigureDto();
-			configureDto.setFindDuplicate(abisInsertService.getDuplicate());
-			return new ResponseEntity<>(configureDto, HttpStatus.OK);
-		} catch (RuntimeException exp) {
-			logger.error("Exception while getting configuration: "+exp.getMessage());
-			throw exp;
-		}
-	}
-
-	@RequestMapping(value = "configure", method = RequestMethod.POST)
-	@ApiOperation(value = "Configure Request")
-	public ResponseEntity<String> configure(@Valid @RequestBody ConfigureDto ie, BindingResult bd)
-			throws Exception {
-		logger.info("Configure Request");
-		try {
-			abisInsertService.setDuplicate(ie.getFindDuplicate());
-			logger.info("[Configuration updated] overrideFindDuplicate: "+abisInsertService.getDuplicate());
-			return new ResponseEntity<>("Successfully updated the configuration", HttpStatus.OK);
-		} catch (RuntimeException exp) {
-			logger.error("Exception in configure request: "+exp.getMessage());
-			throw exp;
-		}
-	}
 	
 	@RequestMapping(value = "insertrequest", method = RequestMethod.POST)
 	@ApiOperation(value = "Save Insert Request")
@@ -206,7 +128,7 @@ public class ProxyAbisController {
 
 	private ResponseEntity<Object> processIdentityRequest(IdentityRequest ir) {
 		logger.info("Finding duplication for reference ID " + ir.getReferenceId());
-		IdentityResponse res = abisInsertService.findDupication(ir);
+		IdentityResponse res = abisInsertService.findDuplication(ir);
 		return new ResponseEntity<Object>(res, HttpStatus.OK);
 	}
