+++ mock-abis/src/main/java/io/mosip/proxy/abis/dao/ProxyAbisBioDataRepository.java
@@ -18,6 +18,6 @@ public interface ProxyAbisBioDataRepository extends CrudRepository<BiometricData
 	@Query(value = "select distinct(b.*) from Biometric_Data b where b.bio_data in (select bio_data from Biometric_Data where Biometric_Data.reference_id =?1) and b.reference_id <> ?1  and b.reference_id in ?2 order by reference_id asc", nativeQuery = true)
 	public List<BiometricData> fetchDuplicatesForReferenceIdBasedOnGalleryIds(@Param("referenceId") String referecenId,@Param("referenceIds") List<String> referenceIds);
 
-	@Query(value = "select bio_data from Biometric_data where reference_id=?1")
+	@Query(value = "select b.bioData from Biometric_Data b where b.insertEntity=?1")
 	public List<String> fetchBiodata(@Param("referenceId") String referenceId);
 }
