+++ mock-abis/src/main/resources/config.properties
@@ -1,4 +1,4 @@
-secret_url=http://kernel-auth-service/v1/authmanager/authenticate/clientidsecretkey
+secret_url=http://sandbox.mosip.net/v1/authmanager/authenticate/clientidsecretkey
 secret_url.clientnId=mosip-regproc-client
 secret_url.id=string
 secret_url.secretKey=abc123
